delete unused storageKeys.ts, move strings.ts to loc folder
