--- extension/src/constants/storageKeys.ts
@@ -1,1 +1,0 @@
-export const selectedProjectKey = 'selectedProject';
